% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CorrSurface.R
\name{CorrSurface}
\alias{CorrSurface}
\title{CorrSurface}
\usage{
CorrSurface(price, lookback_seq, holddays_seq, Sign = FALSE,
  return_method = c("arithmetic", "log"))
}
\arguments{
\item{price}{an object of time serise representing a price series.}

\item{lookback_seq}{a sequence of lookback length.}

\item{holddays_seq}{a sequence of holddays length.}

\item{Sign}{whether the returns or signs of returns are used to calculate
correlation, default is FALSE.}

\item{return_method}{method of calculating returns.}
}
\value{
a 3_D plot of correlation coefficient on lookback period and holding
  period.
}
\description{
The function
}
\examples{
lookback_seq <- seq(from = 1, to = 200, by = 10)
holddays_seq <- seq(from = 1, to = 100, by = 10)
getSymbols("^GSPC", from = 2010, to = Sys.Date())
price <- GSPC[, 6]
CorrSurface(price, lookback_seq = lookback_seq, holddays_seq = holddays_seq,
Sign = TRUE, return_method = "log")
}

