% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LookbackHoldCorr.R
\name{LookbackHoldCorr}
\alias{LookbackHoldCorr}
\title{LookbackHoldCorr}
\usage{
LookbackHoldCorr(price, lookback, holddays, Sign = FALSE,
  return_method = c("arithmetic", "log"))
}
\arguments{
\item{price}{an object of time serise representing a price series.}

\item{lookback}{length of lookback period.}

\item{holddays}{length of holding period.}

\item{Sign}{whether the returns or signs of returns are used to calculate
correlation, default is FALSE.}

\item{return_method}{method of calculating returns.}
}
\value{
a 1 by 4 matrix with lookback period, holding period, return
  correlation estimate and p value.
}
\description{
The function applies discrete moving window to a price series. Each window is
divided into lookback period and holding period. The returns in the lookback
period and holding period are calculated with each window. The correlation
between the associated p value on a returns in the lookback period and
holding period discrete rolling window basis. The rolling step size is the
minimum of the lookback period and holding period.
}
\examples{
getSymbols("^GSPC", from = 2010, to = Sys.Date())
price <- GSPC[, 6]
LookbackHoldCorr(price, 200, 20, TRUE, "log")
LookbackHoldCorr(price, 200, 20, FALSE, "log")
LookbackHoldCorr(price, 200, 20, TRUE, "arithmetic")
LookbackHoldCorr(price, 200, 20, FALSE, "arithmetic")
}

