% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCorrSurface.R
\name{plot.CorrSurface}
\alias{plot.CorrSurface}
\title{plot.CorrSurface()}
\usage{
\method{plot}{CorrSurface}(x)
}
\arguments{
\item{x}{x is a CorrSurface object returned by function CorrSurface()}
}
\value{
plot a 3 D surface of the correlation against lookback and holding
  period with plotly
}
\description{
plot method for object of class CorrSurface.
}

