% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ScreenPara.R
\name{ScreenPara}
\alias{ScreenPara}
\title{Screen}
\usage{
ScreenPara(x, rho_min = 0.3, p_max = 0.1)
}
\arguments{
\item{x}{a CorrSurface object returned by function CorrSurface()}

\item{rho_min}{the minimum threshold for rho to screen parameter combination}

\item{p_max}{the maximum threshold for p value to screen parameter combination}
}
\value{
a object of class ScreenResult
}
\description{
dfsdfsdfsd
}

